requirements_langchain.txt
# LangChain and dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.0.2

# LLM providers
openai>=1.0.0
anthropic>=0.8.0
tiktoken>=0.5.0

# UI automation dependencies
pyautogui>=0.9.54
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10  # For OCR capabilities

# AWS dependencies (for SageMaker)
boto3>=1.28.0
sagemaker>=2.200.0

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Optional: Local LLM support
# llama-cpp-python>=0.2.0

# Optional: Web automation
# selenium>=4.15.0
# webdriver-manager>=4.0.0
